// src/types/ScrollToComponentContextType.ts
export type ScrollToComponentContextType = {
  ScrollToNavBarRef: any;
  ScrollToNavBarFn: () => void;
  ScrollToHomeRef: any;
  ScrollToHomeFn: () => void;
  ScrollToAboutRef: any;
  ScrollToAboutFn: () => void;
  ScrollToPortfoliosRef: any;
  ScrollToPortfoliosFn: () => void;
  ScrollToSkillsRef: any;
  ScrollToSkillsFn: () => void;
  ScrollToContactRef: any;
  ScrollToContactFn: () => void;
};